perf: Phase 6 — i32 node coords, 14× smaller data
## Change
- Store pre-projected node coords as i32 instead of i64
- World-pixels at zoom 20 fit in i32 (max ~268M < 2^31)
- Saves 8 bytes per node: 56 MB vs 83 MB (was 791 MB XML)

## Final data sizes (New Delhi)
| Format | Size | Reduction |
|---|---|---|
| XML | 791 MB | 1× |
| V1 binary | 128 MB | 6.2× |
| V2 i64 nodes | 83 MB | 9.5× |
| V2 i32 nodes | 56 MB | 14.1× |

Cast to i64 before subtraction in renderer to avoid overflow.

+++ src/render_data.h
@@ -10,11 +10,11 @@
 
 struct RenderData {
     // Pre-projected world-pixel coordinates at reference_zoom.
-    // Indexed by node index (0..N-1). During rendering, project to target zoom
-    // via simple shift: pixel = (wx >> (ref_zoom - target_zoom)).
+    // Indexed by node index (0..N-1). At zoom 20, world-pixels fit in i32 (max ~268M).
+    // During rendering, project via: pixel = ((wx - cx) >> shift) + half_w.
     struct ProjNode {
-        int64_t wx;  // world pixel X at reference_zoom
-        int64_t wy;  // world pixel Y at reference_zoom
+        int32_t wx;  // world pixel X at reference_zoom (fits in i32 at z≤20)
+        int32_t wy;  // world pixel Y at reference_zoom
     };
     std::vector<ProjNode> nodes;
